fix: OGPメタデータを静的レンダリングに修正してSNS共有対応
